Create AddBookViewModel
